Clean / Fix ProfileHeader
